[IMP] l10n_jo/saas~18.5: Update taxes, tax grids and tax report
In the community PR https://github.com/odoo/odoo/pull/219061 ,
the tax tags for all taxes are changed, new taxes are added and
the tax report is re-structured as per the Jordanian Government
guidelines.

This upgrade PR does the following things:
- Some existing taxes that are modified in Jordan Accounting but
that changes are not reflected by "Reload". It makes those changes.
It removes the tax repartition lines of some taxes that was modified.
- Tax tags were changed for all the taxes, so in this PR, all that
tax tags are renamed in the pre-migrate script to match with the new
naming convention.
- In the end-migrate, the Chart of account for Jordan companies is
reloaded so that the newly added taxes in the csv are created.

see: https://github.com/odoo/odoo/pull/219061

task-4822933

closes odoo/upgrade#8263

Related: odoo/odoo#219061
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>