[FIX] {pos_sale_stock_renting,pos_sale}/saas~16.4: create column
Summary: `pos_sale_stock_renting` got `pos_sale` as dependency in 16.4

Details: the dependency was added in 16.0 in stable in
odoo/enterprise#42238 then forward ported to then master 16.4. But it
was later removed in odoo/enterprise#42984 but only up to 16.3

We end up with the new column for `sale_order` from 16.4. We cannot use
the same computation as in 16 (even though the compute method is the
same from 16 to 17) because the column sale_order_origin_id doesn't
exist in a fresh install of `pos_sale`.

Compute direct link in 17.0
https://github.com/odoo/odoo/blob/0329cc1b74a9ee2598087e60b825c04b179a0645/addons/pos_sale/models/sale_order.py#L31-L35

Here we just assume that real-world DBs aren't going to have
`pos_sale_stock_renting` installed without `pos_sale`.