[IMP] l10n_ke_hr_payroll/saas~18.5: add commuter and housing allowances
We add a commuter and housing allowance. The housing allowance is either a fixed
 amount or a percentage of the gross amount.

We also change the max food allowance rule parameter for December 2024.

This also makes some fixes, such as removing the employee's pin, which is a
duplicate of the KRA pin. Lastly, we reorder and rename some salary rules.

Task: 4922304

See odoo/enterprise#89976

closes odoo/upgrade#8278

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>