[IMP] auth_passkey/saas~18.5: security rules should be noupdate
Security rules should be noupdate to allow the user to modify them.
This was an oversight and has been fixed with the introduction of
auth_passkey_portal because the commit responsible also introduces a
new security rule. However due to the security rules being noupdate="0"
we don't necessarily need to set that new rule's noupdate parameter. We
simply set the noupdate="1" post-migrate.

See odoo/odoo#223077

closes odoo/upgrade#8280

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>