[FIX] l10n_mx_hr_payroll/saas~18.5: fix ISR payslip overlaps 2 months
Bug:
When a payslip is overlapping two months, the ISR can be negative and is not
correctly computed.

Cause:
There is an adjustment made at the end of each month and each year. This works
only if the payslip ends on the last day of the month.

Fix:
We remove the adjustment and add an input for adjusting the ISR manually.

Task: 4790249

See odoo/enterprise#92813

closes odoo/upgrade#8284

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>