[FIX] account_reports/saas~18.3: misaligned move date
Fix first closing move misalignment

The first closing move is misaligned because the system
currently uses a fixed interval of -1 month.
In addition, the start date was always offset by one day,
causing it to align with the end of the previous period
instead of the start of the current one.

At the moment, the period of a return is determined by checking
for a previous return: if one exists, the new period starts from
the end of the previous; if not, the system simply goes back x
months from the closing date.
This approach is fragile, as it doesn’t properly handle irregularities
in the calendar, such as the varying length of February.

The solution is to rely on the same method used to generate
the return itself, ensuring consistency, correct alignment,
and support for complex calendar cases.

opw-5029344

closes odoo/upgrade#8288

Enterprise: odoo/enterprise#93100
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>