[FIX] account_accountant/saas~18.3: avoid reco trigger
We do not want a reco model to auto trigger when we apply them. The main
purpose of #7974 was to ensure the reco model was linked to the right
records. As a side effect we got that the automatic reco models trigger
posting or removal of data, this affects the balances of some accounts.
That breaks the invariant check for account balances.

To avoid the issue we temporarily remove the auto trigger from all reco
models when applying them.

closes odoo/upgrade#8293

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>