[FIX] {base_vat,l10n_uy}/17.0: compute column
`l10n_uy` depends on `l10n_latam_base` since 16.5. The latter depends on
`base_vat`. When `base_vat` is being installed it will compute the new
filed `vies_valid`. This column is already computed in
`base_vat/saas~16.3.1.0/pre-migrate.py` we only need to ensure it is
filled as well in 17.0 as soon as we detect that the module is being
installed.