[IMP] account_iso20022/saas~18.5: move end_to_end_id
The field existed in account_online_payment but was never used. In the
main task, the goal is to use this field for reconciliation. The field
was moved and renamed in account_iso20022 and computed correctly using
the hex value of a UUID4.

Enterprise PR: https://github.com/odoo/enterprise/pull/92300

task-4210735

closes odoo/upgrade#8310

Related: odoo/enterprise#92300
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>