[FIX] account_reports/saas~18.5: use indentation instead of return
The pre script checks for the existence of a table to
perform some operations and returns if not found.
Using return in a generic script leads to undesired
side effects like skipping everything that is added
later at the bottom of the file.
The field rename in the end of the file needs to be
executed in all cases.

closes odoo/upgrade#8315

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>