[IMP] appointment/saas~18.5: adapt allday slot end datetime
Make sure that the end datetime of allday slots match the correct
consensus with the possibilities to have allday slots spanning
multiple days. As we expect only to have allday slot for a single
day, we just set the end_datetime to the start_datetime.

ENT PR: odoo/enterprise#81734

task-2822534

closes odoo/upgrade#8323

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>