[IMP] account_reports/saas~18.5: change `result` selection
The related PR removes the need to check `bypassed` and `result` in
order to determine if the check is ok or not.
Now `result` holds the state of the check.  `refresh_result` is set to
false when a user change the state of a check, so that the state can no
longer be changed automatically.

The selection of `result` of `account.return.check` is changed from
['failure', 'manual', 'success'] to ['anomaly', 'todo', 'reviewed',
'supervised'].

The field `bypassed` is removed, `refresh_result` is added and is set
to the inverse value of `bypassed`.

enterprise: https://github.com/odoo/enterprise/pull/92086

task-5013703

closes odoo/upgrade#8332

Related: odoo/enterprise#92086
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>