[FIX] hr_payroll/0.0.0: noupdate for structs with countries
If a user modified a standard payroll structure
and added a country id, the record would be
reverted to have no countries when the module
data loads. This has consequences also on the
related salary rules, where some rules that
used to appear on a payroll report would stop
appearing, and the rules' artifical field
on the report are removed.
Removing the artificial field is what is
causing the upgrades to fail, but all the
undesirable consequences can be avoided
by setting the user modified structures to
noupdate.

see https://github.com/odoo/enterprise/pull/83749

closes odoo/upgrade#8339

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>