[FIX] {crm,project}/saas~16.4: Replace view_in_kanban in the properties
The 'View in kanban' option was not available in 16 for crm leads and project
tasks, but before it was patched out [^1], it was possible to use that option.
This option was restored in 17 when it got replaced[^2] with 'View in cards'.

Some databases that installed the modules before the fix, might have the wrong
key (view_in_kanban) in their properties, so it's necessary to change them to
the new key to avoid errors when manipulating the properties in the ugpraded
database.[^3]

opw-5017430
upg-3062824

```
ValueError: Some key are not allowed for a properties definition [view_in_kanban].
```

[^1]: https://github.com/odoo/odoo/commit/90b603c28234b587739f7e16da70fbc5e6d85a60
[^2]: https://github.com/odoo/odoo/commit/f7feeedc1530e9cc896dbb758594c320a073a352
[^3]: https://github.com/odoo/upgrade/commit/73e550eb1b38faa694a61e7d70c2b00a9cc4eefe

closes odoo/upgrade#8353

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>