[FIX] product/saas~18.5: move demo xmlid creation
The xmlid for variant product_product_acoustic_bloc_screens_black
was only added in 'saas~18.5', the error was still visible when
migrating from 'saas~18.4'. This fix moves the xmlids creation
to the correct version, while also checking if the variant
itself needs creation for the case of coming from lower versions.

closes odoo/upgrade#8356

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>