[FIX] appointment/saas~18.4.1.3: Fix mail_template translation
Steps to reproduce the error:

1. Create a demo database in version saas~18.2 or 18.3.
2. Install the Appointment app.
3. Add any language except English.
4. Migrate the demo database to saas~18.4.
5. After migration, go to the Appointment module and book an appointment.
6. The error appears.

Root Cause:

The error occurs when booking an appointment in a language other than English
because the resource_manage_capacity and resource_total_capacity_reserved fields
have been renamed in saas~18.4. Here is the link:
https://github.com/odoo/upgrade/blob/efe6b3597a610f81d903010899c1cc25be3c3021/migrations/appointment/saas~18.4.1.3/pre-migrate.py
and these changes are not reflected in the body_html of the mail template
in other languages, and the translations mentioned in the .po files for other
languages are not updated.

closes odoo/upgrade#8389

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>