[FIX] website_sale/saas~19.1: migrate to global GMC flag
In [^1], a decision was made to introduce a new global group-feature
flag to control the visibility of related menus. However, this flag was
not enabled during the upgrade to version 19 (18.5).

Additionally, in [^2], it was decided to merge the GMC feature flags
into the global group-feature flag, removing the website-specific flag.
This commit updates the GMC feature flag migration to reflect this
decision.

[^1]: https://github.com/odoo/odoo/pull/224889
[^2]: https://github.com/odoo/odoo/pull/227629

closes odoo/upgrade#8468

Related: odoo/odoo#227629
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>