[FIX] base/saas~16.3: ignore ORM changes
`spreadsheet` gets the dependency `portal` in 16.3. `portal` depends on
`mail`, hence the latter will be installed. In real world databases we
don't expect `spreadsheet` module to be used without `mail` **and**
having lots of users/partners. Thus in order to keep the scripts here
simpler, we opt to ignore this case.