[FIX] hr_holidays/saas~16.5: cannot use ratio for hours
Since the final operation is `hours * ratio`, and the `hours` refer to
the total number of hours in the leave, not the hours per day. Thus the
previous ratio was wrong. Even if we fix it to be `days / hours *
leave.resource_calendar_id.hours_per_day` instead of `days`, it fails
for calendars where the number of hours isn't the same all days. A test
is included to cover this case.

The solution is then to actually use the compute method. We can restore
the number of days to keep any manual modification.

upg-3138728

closes odoo/upgrade#8495

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>