[IMP] gmail: improve the email logging
Purpose
=======
Before the subject, email from, etc were added in the body of the
email when we log it in a record. Now, those values are properly written
in the fields of the `mail.message`, and so we need to send them to the
Odoo endpoint.

Parse all the contacts in the email TO, CC,... to prepare the
following commit.

Before, if we wait some time before logging the email on a record,
an error could be raised because the token we received to get the
information expired. To solve that issue, we parse and save then in
the state when the addin is loaded.

Task-4727609