[IMP] gmail: remove the enrichment feature from the addon
Purpose
=======
Then enrichment feature of the addon made it slow and hard to use in
practice, so we simplify it, and we keep only the core feature, which
is logging the email on the records and viewing information about the
contact.

If a conversation contains many contacts, before it took the first
one. We can now choose the contact we want to open.

Allow searching any records, not only partners.

We fixed some UI issues than were introduced by
update of the Gmail API.

Task-4727609