[IMP] gmail: move from appscript to HTTP addin
Purpose
=======
The current addin is slow, because of how appscript work (Google host
a web server, with our app running, that server restart at each call,
the language itself is synchronous...).

In alternative way of creating addin, is to host the application ourselves.
It's faster because we can keep the server alive, and we can make the
HTTP calls in parallel (to the Gmail API and to the Odoo database).
> https://developers.google.com/workspace/add-ons/guides/alternate-runtimes

We choose to use node with express, because it's asynchronous, and we
will make a lot of request (to the Odoo database, to the Gmail API, etc).

Also, most of the code can be re-used (the core logic).

The main changes compared to appscript are
- we need to store the user settings ourselves (and we use psql for that)
- so the translations and the email logging state is moved to the user
  (it's considered as "User Settings / Information").
- to build the view, our endpoints need to return JSON, so we created
  some helper classes similar to the one use by appscript (no official
  library exists for that at the moment, for typescript)

The application has been kept as similar as the one using appscript as
possible (the execution of action, with the state of the current page
is sent in the response).

Task-4727609