[IMP] gmail: serve image file instead of encoding in base 64
Purpose
=======
When we used appscript, we didn't have a web server to store the image
(and to do the processing to translate text inside the SVG), and so
we base 64 encoded them. Now that we moved the HTTP addin, we have a
web server, and we can serve the files (and translate them on the fly),
so that the images are cached by the browser, and we don't download
them each time we view a new card.

When Gmail shows images, it doesn't add our URL in the dom, instead
it fetches the image, and store it in `googleusercontent.com`. But,
it won't fetch the images if they come from a ngrok domain name, so
now, we need a "real" domain name for development.

Task-4727609