[IMP] outlook: remove the enrichment feature from the addon
Purpose
=======
Then enrichment feature of the addon made it slow and hard to use in
practice, so we simplify it, and we keep only the core feature, which
is logging the email on the records and viewing information about the
contact.

If a conversation contains many contacts, before it took the first
one. We can now choose the contact we want to open.

Allow searching any records, not only partners.

Refactoring
===========
The library we used is obsolete ("UI Fabric"), and so we took advantage
of that task to re-write it using the new library "fluentui".

Task-5152391